stop eagerly filtering dependency closure
Using builtins.genericClosure is working against trying to support
various things like development shells, tests and the like. Hence, this
commit drops fetch-sources which only downloads the sources for an
explicit dependency closure in favor of defining downloads for all
packages and relying on the laziness of nix to only download what's
necessary.